+++ packages/components/modules/content-feed/common/index.ts
@@ -1,3 +1,6 @@
 // exports common content-feed code
 
 export * from './graphql/mutations/ContentPostCreate'
+export * from './graphql/fragments/ContentPosts'
+export * from './graphql/queries/ContentPosts'
+export * from './graphql/queries/ContentPostDetail'